+++ src/dd_sample_definition.cpp
@@ -169,6 +169,38 @@ void SampleInfo::gettotalreadnum(const std::string &filename, const std::vector<
 #endif
 }
 
+void vSampleInfo::addSampleInfo(const std::string &str,
+				const std::vector<chrsize> &gt,
+				const WigType iftype)
+{
+  int32_t binsize(0);
+  std::vector<std::string> v;
+  ParseLine(v, str, ',');
+
+  if(v.size() >8) {
+    PRINTERR_AND_EXIT("error: sample std::string has ',' more than 8: " << str);
+  }
+  if(v[0] == "") {
+    PRINTERR_AND_EXIT("please specify ChIP sample: " << str);
+  }
+  isFile(v[0]);
+
+  if(v.size() >4 && v[4] != "") {
+    try { binsize = stoi(v[4]); }
+    catch (...) { std::cerr << "Warning: invalid binsize " << v[4] << "." << std::endl; }
+  }
+
+  // ChIP sample
+  if(!Exists(v[0])) vsinfo[v[0]] = SampleInfo(v[0], gt, binsize, iftype);
+  if(vsinfo[v[0]].getbinsize() <= 0) PRINTERR_AND_EXIT("please specify binsize.\n");
+
+  // Input sample
+  if(v.size() >=2 && v[1] != "") {
+    if(!Exists(v[1])) vsinfo[v[1]] = SampleInfo(v[1], gt, binsize, iftype);
+    if(vsinfo[v[0]].getbinsize() != vsinfo[v[1]].getbinsize()) PRINTERR_AND_EXIT("binsize of ChIP and Input should be same. " << str);
+  }
+}
+
 
 SamplePairEach::SamplePairEach(const std::string &str, const vSampleInfo &vsinfo):
   binsize(0),
@@ -181,7 +213,7 @@ SamplePairEach::SamplePairEach(const std::string &str, const vSampleInfo &vsinfo
      6:scale_tag   7:scale_ratio   8:scale_pvalue */
   if (v[0] != "") argvChIP = v[0];
   if (v.size() >=2 && v[1] != "") argvInput = v[1];
-  if (v.size() >=3 && v[2] != "") label     = v[2];
+  if (v.size() >=3 && v[2] != "") label     = v[2]; else label = v[0];
   if (v.size() >=4 && v[3] != "") peak_argv = v[3];
   if (peak_argv != "") vbedregions = parseBed_Hash<bed>(peak_argv);
   binsize = vsinfo.getbinsize(argvChIP);